// src/app/api/models/country.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Detalles del Country a crear.
 */
export interface Country {

  /**
   * Identificador unico
   */
  id?: number;

  /**
   * ISO code of Country
   */
  iso: string;

  /**
   * ISO3 of Country
   */
  iso3?: string;

  /**
   * Name of Country
   */
  name: string;

  /**
   * Nice name of Country
   */
  niceName: string;

  /**
   * Num Code of Country
   */
  numCode?: number;

  /**
   * Phone code of the Country
   */
  phoneCode: number;
}